refact: Moved mdZipFile function to MdDownloadService layer
